register_serv/requirements.txt
pip==22.1.2
click==8.1.3
setuptools==62.6.0
Jinja2==3.1.2
Flask==2.1.2
Flask-Migrate==3.1.0
WTForms==3.0.1
Flask-SQLAlchemy==2.5.1
itsdangerous==2.1.2
Werkzeug==2.1.2
MarkupSafe==2.1.1
alembic==1.8.0
passlib==1.7.4
Flask-Admin==1.6.0
SQLAlchemy==1.4.39
blinker==1.4.0
Mako==1.2.0
greenlet==1.1.2
python2-secrets==1.0.5
PyMySQL==1.0.2
Flask-WTF==1.0.1
wheel==0.37.1
python-dotenv==0.20.0
Flask-Login==0.6.1
config==0.5.1
colorama==0.4.5
cryptography==37.0.2
pytest==7.1.2

